Moved loopback-fw over to the RTIC version, on the dongle.
Tested on macOS 15, Windows 10 and Debian 12.
