--- nrf52-code/loopback-fw-rtic/.cargo/config.toml
@@ -1,23 +1,0 @@
-[target.thumbv7em-none-eabihf]
-# set custom cargo runner to flash & run on embedded target when we call `cargo run`
-# for more information, check out https://github.com/probe-rs
-runner = [
-  "probe-rs",
-  "run",
-  "--chip",
-  "nRF52840_xxAA",
-  "--allow-erase-all",
-  "--log-format",
-  "{t} {[{L}]%bold} {s} {({c:bold} {fff}:{l:1})%dimmed}"
-]
-# Or use "{[{L}]%bold} {s} {({c:bold} {fff}:{l:1})%dimmed}" to disable timestamps
-
-linker = "flip-link" # adds stack overflow protection
-rustflags = [
-  "-C", "link-arg=-Tlink.x", # use the cortex-m-rt linker script
-  "-C", "link-arg=-Tdefmt.x", # defmt support
-]
-
-[build]
-# cross-compile to this target
-target = "thumbv7em-none-eabihf" # = ARM Cortex-M4 with FPU
