--- nrf52-code/loopback-fw-rtic/.gitignore
@@ -1,1 +1,0 @@
-target
