+++ nrf52-code/loopback-fw/Cargo.toml
@@ -7,33 +7,12 @@ version = "0.0.0"
 
 [dependencies]
 consts = { path = "../consts" }
-cortex-m = {version = "0.7.7", features = ["critical-section-single-core"]}
-cortex-m-rt = "0.7.5"
-critical-section = "1.2.0"
+defmt = "0.3.10"
+defmt-rtt = "0.4.1"
 dongle = { path = "../boards/dongle" }
-embedded-hal = "1.0"
-usb-device = "0.3"
-usbd-hid = "0.8"
-usbd-serial = "0.2"
-
-[build-dependencies]
-rand = "0.8.5"
-
-# optimize code in both profiles
-[profile.dev]
-codegen-units = 1
-debug = 2
-debug-assertions = true # !
-incremental = false
-lto = "fat"
-opt-level = 'z' # !
-overflow-checks = false
-
-[profile.release]
-codegen-units = 1
-debug = 2
-debug-assertions = false
-incremental = false
-lto = "fat"
-opt-level = 3
-overflow-checks = false
+heapless = { version = "0.8.0", features = ["defmt-03"] }
+rtic = { version = "2.1.2", features = ["thumbv7-backend"] }
+rtic-monotonics = { version = "2.0.3", features = ["cortex-m-systick"] }
+usb-device = { version = "0.3.2", features = ["defmt"] }
+usbd-hid = { version = "0.8.2", features = ["defmt"] }
+usbd-serial = "0.2.2"
